my sixth work

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int f;
 	float e;
 
-	printf("Size of a char:%ld byte(S)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %ld byte(S)\n", (unsigned long)sizeof(b));
-	printf("Size of a long: %ld byte(S)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %ld byte(S)\n", (unsigned long)sizeof(f));
-	printf("Size of a float:%ld byte(S)\n", (unsigned long)sizeof(e));
+	printf("Size of a char:%ld byte(S)\n", sizeof(a));
+	printf("Size of an int: %ld byte(S)\n", sizeof(b));
+	printf("Size of a long: %ld byte(S)\n", sizeof(c));
+	printf("Size of a long long int: %ld byte(S)\n", sizeof(f));
+	printf("Size of a float:%ld byte(S)\n", sizeof(e));
 	return (0);
 }
